Change hidden scroll bar buttons to collapsed
When scroll bars are not expanded there are gaps reserved for the
scroll bar buttons that appear when the scroll bar expands. This is
misleading though because the user thinks he can scroll even further
as there are no buttons visible and the thumb did not reach the end.
This commit removes the gaps and adds animation to make space
for the buttons smoothly when they appear.